Moved SaveButton into its own file
